Refactor LTX pipelines around shared runtime
