+++ vllm_omni/diffusion/models/ltx2/__init__.py
@@ -1,14 +1,16 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 
+from vllm_omni.diffusion.models.ltx2.ltx2_components import (
+    create_transformer_from_config,
+    load_transformer_config,
+)
 from vllm_omni.diffusion.models.ltx2.ltx2_transformer import LTX2VideoTransformer3DModel
 from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import (
     LTX2Pipeline,
     LTX2T2VDMD2Pipeline,
     LTX2TwoStagesPipeline,
-    create_transformer_from_config,
     get_ltx2_post_process_func,
-    load_transformer_config,
 )
 from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3 import LTX23Pipeline
 from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
